+++ src/manifold_optimizers/gradient_optimizer.jl
@@ -51,8 +51,7 @@ struct GradientCache{T, MT, VT, ST} <: OptimizerCache{T}
     section::ST
 end
 
-function GradientCache(
-        x::OptimizerSolution{T}, g::AT, δ::AT, Δg::AT) where {
+function GradientCache(x::OptimizerSolution{T}, g::AT, δ::AT, Δg::AT) where {
         T, AT <: GradientStorage{T}}
     sec = GlobalSection(_copy(x))
     g̃ = _similar(g)
@@ -60,8 +59,7 @@ function GradientCache(
     GradientCache{T, typeof(x), typeof(g), typeof(sec)}(x, g, δ, Δg, g̃, Ref(false), sec)
 end
 
-function GradientCache(
-        x::OptimizerSolution{T}, g::AT, δ::AT) where {
+function GradientCache(x::OptimizerSolution{T}, g::AT, δ::AT) where {
         T, AT <: GradientStorage{T}}
     Δg = _similar(g)
     _fill!(Δg, T(NaN))
@@ -90,8 +88,7 @@ latest_gradient(cache::GradientCache) = cache.g̃
 function refresh_latest_gradient!(cache::GradientCache, g::Gradient)
     _refresh_latest_gradient!(cache, g)
 end
-function latest_gradient_is_current(
-        cache::GradientCache, state::OptimizerState, x::OptimizerSolution)
+function latest_gradient_is_current(cache::GradientCache, state::OptimizerState, x::OptimizerSolution)
     _latest_gradient_is_current(cache, state, x)
 end
 invalidate_latest_gradient!(cache::GradientCache) = _invalidate_latest_gradient!(cache)